move submission and onfinished to orchestrated selfie capture viewmodel.
